Added an outline for edges.
Setting the stroke to a different value than the fill had weird effects
due to the way paths work. Used a filter instead. The problem is, the
filter does cause a bit of a performance hit. This is probably
something that should be turned into an optional feature toggled by a
button that is turned off by default.